Break up sample into separate files
